+++ lib/punchblock/connection/xmpp.rb
@@ -61,7 +61,7 @@ module Punchblock
         command.mixer_name    ||= options[:mixer_name]
         command.component_id  ||= options[:component_id]
         create_iq(jid_for_command(command)).tap do |iq|
-          pb_logger.debug "Sending IQ ID #{iq.id} #{command.inspect} to #{jid}"
+          pb_logger.debug "Sending IQ ID #{iq.id} #{command.inspect} to #{iq.to}"
           iq << command
         end
       end
@@ -78,11 +78,6 @@ module Punchblock
           EM.run { client.run }
         rescue Blather::Stream::ConnectionFailed, Blather::Stream::ConnectionTimeout => e
           raise DisconnectedError.new(e.class.to_s, e.message)
-        rescue => e
-          # Preserve Punchblock native exceptions
-          raise e if e.class.to_s =~ /^Punchblock/
-          # Wrap everything else
-          raise ProtocolError.new(e.class.to_s, e.message)
         end
       end