+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # develop
 
+# v0.9.2 - 2012-02-18
+  * Feature: Asterisk calls receiving media commands are implicitly answered
+  * Bugfix: Unrenderable output documents on Asterisk should return a sensible error
+  * Bugfix: Log the target of commands correctly
+  * Bugfix: Do not wrap exceptions in ProtocolError
+
 # v0.9.1 - 2012-01-30
   * Bugfix: Closing an disconnected XMPP connection is a no-op